Update main.c
fixing data type back to float

+++ main.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-int subtract(float num1,float num2);
+float subtract(float num1,float num2);
 float add(float num1,float num2);
 float multiplication(float num1, float num2);
 float division(float num1, float num2);
